Task13 (#14)
* task13

* task13

* task13

* task13

* task13

+++ src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -20,7 +20,7 @@ public class FileBasedDatabase implements Database {
   private final JsonService jsonService;
 
   @Override
-  public int save(Invoice invoice) {
+  public long save(Invoice invoice) {
     try {
       invoice.setId(idProvider.getNextIdAndIncrement());
       fileService.appendLineToFile(dbPath, jsonService.toJson(invoice));
@@ -32,7 +32,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> getById(int id) {
+  public Optional<Invoice> getById(long id) {
     try {
       return fileService.readAllLines(dbPath)
           .stream()
@@ -57,7 +57,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+  public Optional<Invoice> update(long id, Invoice updatedInvoice) {
     try {
       List<String> allInvoices = fileService.readAllLines(dbPath);
       var invoicesWithoutInvoiceWithGivenId = allInvoices
@@ -80,7 +80,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> delete(int id) {
+  public Optional<Invoice> delete(long id) {
     try {
       var allInvoices = fileService.readAllLines(dbPath);
 
@@ -100,7 +100,7 @@ public class FileBasedDatabase implements Database {
     }
   }
 
-  private boolean containsId(String line, int id) {
+  private boolean containsId(String line, long id) {
     return line.contains("{\"id\":" + id + ",\"number\"");
   }
 }
